Rename backpack connector to follow naming conventions
