+++ hummingbot/connector/exchange/backpack/backpack_api_order_book_data_source.py
@@ -2,7 +2,7 @@ import asyncio
 import time
 from typing import TYPE_CHECKING, Any, Dict, List, Optional
 
-from hummingbot.connector.exchange.backpack_exchange import backpack_constants as CONSTANTS
+from hummingbot.connector.exchange.backpack import backpack_constants as CONSTANTS
 from hummingbot.core.data_type.order_book_message import OrderBookMessage, OrderBookMessageType
 from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
 from hummingbot.core.web_assistant.connections.data_types import RESTMethod, WSJSONRequest
