+++ hummingbot/connector/exchange/backpack/backpack_constants.py
